Remove AddSiteProvider and move state management to useAddSite hook
