+++ src/components/tests/main-sidebar.test.tsx
@@ -1,7 +1,6 @@
 import { render, act, screen } from '@testing-library/react';
 import { userEvent } from '@testing-library/user-event';
 import { Provider } from 'react-redux';
-import { AddSiteProvider } from 'src/components/add-site-provider';
 import MainSidebar from 'src/components/main-sidebar';
 import { SyncSitesProvider } from 'src/hooks/sync-sites';
 import { useAuth } from 'src/hooks/use-auth';
@@ -102,9 +101,7 @@ const renderWithProvider = ( children: React.ReactElement ) => {
 	return render(
 		<Provider store={ store }>
 			<ContentTabsProvider>
-				<SyncSitesProvider>
-					<AddSiteProvider>{ children }</AddSiteProvider>
-				</SyncSitesProvider>
+				<SyncSitesProvider>{ children }</SyncSitesProvider>
 			</ContentTabsProvider>
 		</Provider>
 	);
